TestDatabase: match new Student interface
Student now has separate firstName and lastName fields. That means a few
things have to change: how Students are added & how we search for them
by name.

Also, accidentally missed the old unused methods which already got a
list of IDs for a name. Will add back testId in some form later, but
studentId needed to be changed because of the changes in Student.